+++ SOLID/Ex7/src/AttendanceReader.java
@@ -1,0 +1,3 @@
+public interface AttendanceReader {
+    int scanAttendance();
+}