+++ SOLID/Ex7/src/PowerSwitch.java
@@ -1,0 +1,4 @@
+public interface PowerSwitch {
+    void powerOn();
+    void powerOff();
+}