+++ SOLID/Ex8/src/events.java
@@ -1,0 +1,5 @@
+public interface events {
+
+        void createEvent(String name, double budget);
+    int getEventsCount();
+}