+++ SOLID/Ex8/src/finance.java
@@ -1,0 +1,5 @@
+public interface finance {
+
+ void addIncome(double amt, String note);
+    void addExpense(double amt, String note);
+}