+++ SOLID/Ex8/src/minutes.java
@@ -1,0 +1,5 @@
+public interface minutes {
+
+        void addMinutes(String text);
+
+}